Safari-609.1.11.1/LayoutTests/accessibility/gtk/spans-expected.txt

first block span
inline span
second block span

first block span
inline span
second block span

first block span
inline span
second block span
first block span
inline span
second block span
This tests that both block spans and inline spans get folded into the parent object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS element.role is 'AXRole: AXHeading'
PASS element.childrenCount is 0
PASS element.role is 'AXRole: AXHeading'
PASS element.childrenCount is 1
PASS link.role is 'AXRole: AXLink'
PASS link.childrenCount is 0
PASS element.role is 'AXRole: AXList'
PASS element.childrenCount is 2
PASS item.role is 'AXRole: AXListItem'
PASS item.childrenCount is 0
PASS item.role is 'AXRole: AXListItem'
PASS item.childrenCount is 1
PASS link.role is 'AXRole: AXLink'
PASS link.childrenCount is 0
PASS successfullyParsed is true

TEST COMPLETE
